Add answer to Exercise 1.14 and 1.15
1.14 Draw the tree illustrating the process generated by the
count-change procedure of section 1.2.2 in making change for 11 cents.

1.15 The sine of an angle(specified in radians) ban be computed by
making use of the approximation sin x ≈ x if x is sufficiently small,
an the trigonometric identity sin x = 3(sin x/3) - 4(sin x/3)^3 to
reduce the size of the argument of sin.
a. How many times is the procedure p applied when (sine 12.15) is
evaluated?
b. What is the order of growth in space and number of steps(as a
function of a) used by the process generated by the sine procedure when
(sine a) is evaluated?