Add answer to Exercise 1.38
And modify the implementation of Exercise 1.37,
In recursive version,  it should start from 1 to k, instead of k to 1.
Since in Exercise 1.37 all the d will be 1, so there is no difference
between from k to 1 and from 1 to k. However in euler-expansion, it
requires strictly follow the formula, so it should starts from 1 to k.

And in iterative version, we have already compute the value of k, at
the beginning of the iteration, so we need to deduct k by one, as well
as the ending condition should be greater than zero.